requirements.txt
playwright>=1.42.0
langchain>=0.1.14
langgraph>=0.0.40
python-dotenv>=1.0.1
gspread>=6.0.2
google-auth>=2.28.1
tenacity>=8.2.3
pydantic>=2.6.3
pandas>=2.1.4


